+++ src/eBayEnterprise/RetailOrderManagement/Payload/Inventory/AbstractQuantityMessage.php
@@ -1,0 +1,72 @@
+<?php
+/**
+ * Copyright (c) 2013-2014 eBay Enterprise, Inc.
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * @copyright   Copyright (c) 2013-2014 eBay Enterprise, Inc. (http://www.ebayenterprise.com/)
+ * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+
+namespace eBayEnterprise\RetailOrderManagement\Payload\Inventory;
+
+use eBayEnterprise\RetailOrderManagement\Payload\IPayload;
+use eBayEnterprise\RetailOrderManagement\Payload\IPayloadMap;
+use eBayEnterprise\RetailOrderManagement\Payload\ISchemaValidator;
+use eBayEnterprise\RetailOrderManagement\Payload\IValidatorIterator;
+use eBayEnterprise\RetailOrderManagement\Payload\PayloadFactory;
+use eBayEnterprise\RetailOrderManagement\Payload\TTopLevelPayload;
+use Psr\Log\LoggerInterface;
+
+abstract class AbstractQuantityMessage
+{
+    use TTopLevelPayload, TQuantityItemContainer;
+
+    /**
+     * @param IValidatorIterator
+     * @param ISchemaValidator
+     * @param IPayloadMap
+     * @param LoggerInterface
+     * @param IPayload
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     */
+    public function __construct(
+        IValidatorIterator $validators,
+        ISchemaValidator $schemaValidator,
+        IPayloadMap $payloadMap,
+        LoggerInterface $logger,
+        IPayload $parentPayload = null
+    ) {
+        $this->logger = $logger;
+        $this->validators = $validators;
+        $this->schemaValidator = $schemaValidator;
+        $this->payloadMap = $payloadMap;
+        $this->parentPayload = $parentPayload;
+        $this->payloadFactory = new PayloadFactory;
+
+        $this->quantityItems = $this->buildPayloadForInterface(
+            $this->getQuantityItemIterableInterface()
+        );
+
+        $this->subpayloadExtractionPaths = [
+            'quantityItems' => '/',
+        ];
+    }
+
+    protected function serializeContents()
+    {
+        return $this->getQuantityItems()->serialize();
+    }
+
+    /**
+     * Get the interface required for the quantity item iterable.
+     *
+     * @return string
+     */
+    abstract protected function getQuantityItemIterableInterface();
+}