+++ src/eBayEnterprise/RetailOrderManagement/Payload/Inventory/IItem.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * Copyright (c) 2013-2014 eBay Enterprise, Inc.
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * @copyright   Copyright (c) 2013-2014 eBay Enterprise, Inc. (http://www.ebayenterprise.com/)
+ * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+
+namespace eBayEnterprise\RetailOrderManagement\Payload\Inventory;
+
+use eBayEnterprise\RetailOrderManagement\Payload\IPayload;
+use eBayEnterprise\RetailOrderManagement\Payload\IPayload\Payment\IShippingAddress;
+
+/**
+ * Order line item in an inventory API request.
+ */
+interface IItem extends IPayload
+{
+    const XML_NS = 'http://api.gsicommerce.com/schema/checkout/1.0';
+
+    /**
+     * Identifier for an order item.
+     *
+     * restrictions: 1 <= length <= 39, unique in request
+     * @return string
+     */
+    public function getLineId();
+
+    /**
+     * @param string
+     * @return self
+     */
+    public function setLineId($id);
+
+    /**
+     * identifier for an inventoriable product. a.k.a. SKU
+     *
+     * restrictions: length <= 20
+     * @return string
+     */
+    public function getItemId();
+
+    /**
+     * @param string
+     * @return self
+     */
+    public function setItemId($id);
+}