+++ src/eBayEnterprise/RetailOrderManagement/Payload/Inventory/QuantityItem.php
@@ -1,0 +1,76 @@
+<?php
+/**
+ * Copyright (c) 2013-2014 eBay Enterprise, Inc.
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * @copyright   Copyright (c) 2013-2014 eBay Enterprise, Inc. (http://www.ebayenterprise.com/)
+ * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+
+namespace eBayEnterprise\RetailOrderManagement\Payload\Inventory;
+
+use eBayEnterprise\RetailOrderManagement\Payload\IPayload;
+use eBayEnterprise\RetailOrderManagement\Payload\IPayloadMap;
+use eBayEnterprise\RetailOrderManagement\Payload\ISchemaValidator;
+use eBayEnterprise\RetailOrderManagement\Payload\IValidatorIterator;
+use eBayEnterprise\RetailOrderManagement\Payload\TPayload;
+use Psr\Log\LoggerInterface;
+
+class QuantityItem implements IQuantityItem
+{
+    use TItem, TPayload;
+
+    /**
+     * @param IValidatorIterator
+     * @param ISchemaValidator
+     * @param IPayloadMap
+     * @param LoggerInterface
+     * @param IPayload
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     */
+    public function __construct(
+        IValidatorIterator $validators,
+        ISchemaValidator $schemaValidator,
+        IPayloadMap $payloadMap,
+        LoggerInterface $logger,
+        IPayload $parentPayload = null
+    ) {
+        $this->logger = $logger;
+        $this->validators = $validators;
+        $this->parentPayload = $parentPayload;
+
+        $this->extractionPaths = [
+            'itemId' => 'string(@itemId)',
+            'id' => 'string(@lineId)',
+        ];
+    }
+
+    protected function getRootNodeName()
+    {
+        return static::ROOT_NODE;
+    }
+
+    protected function getXmlNamespace()
+    {
+        return static::XML_NS;
+    }
+
+    protected function serializeContents()
+    {
+        return '';
+    }
+
+    protected function getRootAttributes()
+    {
+        return [
+            'lineId' => $this->getId(),
+            'itemId' => $this->getItemId(),
+        ];
+    }
+}