+++ src/eBayEnterprise/RetailOrderManagement/Payload/Inventory/QuantityItemIterable.php
@@ -1,0 +1,88 @@
+<?php
+/**
+ * Copyright (c) 2013-2014 eBay Enterprise, Inc.
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * @copyright   Copyright (c) 2013-2014 eBay Enterprise, Inc. (http://www.ebayenterprise.com/)
+ * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+
+namespace eBayEnterprise\RetailOrderManagement\Payload\Inventory;
+
+use eBayEnterprise\RetailOrderManagement\Payload\IPayload;
+use eBayEnterprise\RetailOrderManagement\Payload\IPayloadMap;
+use eBayEnterprise\RetailOrderManagement\Payload\ISchemaValidator;
+use eBayEnterprise\RetailOrderManagement\Payload\IValidatorIterator;
+use eBayEnterprise\RetailOrderManagement\Payload\PayloadFactory;
+use eBayEnterprise\RetailOrderManagement\Payload\TIterablePayload;
+use Psr\Log\LoggerInterface;
+use SPLObjectStorage;
+
+class QuantityItemIterable extends SPLObjectStorage implements IQuantityItemIterable
+{
+    use TIterablePayload;
+
+    const SUBPAYLOAD_XPATH = '*';
+
+    /**
+     * @param IValidatorIterator
+     * @param ISchemaValidator
+     * @param IPayloadMap
+     * @param LoggerInterface
+     * @param IPayload
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     */
+    public function __construct(
+        IValidatorIterator $validators,
+        ISchemaValidator $schemaValidator,
+        IPayloadMap $payloadMap,
+        LoggerInterface $logger,
+        IPayload $parentPayload = null
+    ) {
+        $this->logger = $logger;
+        $this->validators = $validators;
+        $this->payloadMap = $payloadMap;
+        $this->parentPayload = $parentPayload;
+        $this->payloadFactory = new PayloadFactory;
+
+        // This payload does not have a root node, it's elements
+        // are located directly inside the parent payload.
+        $this->buildRootNode = false;
+    }
+
+    /**
+     * Get a new, empty quantity item.
+     *
+     * @return IQuantityItem
+     */
+    public function getEmptyQuantityItem()
+    {
+        return $this->buildPayloadForInterface(static::QUANTITY_ITEM_INTERFACE);
+    }
+
+    protected function getNewSubpayload()
+    {
+        return $this->getEmptyQuantityItem();
+    }
+
+    protected function getSubpayloadXPath()
+    {
+        return self::SUBPAYLOAD_XPATH;
+    }
+
+    protected function getRootNodeName()
+    {
+        return static::ROOT_NODE;
+    }
+
+    protected function getXmlNamespace()
+    {
+        return self::XML_NS;
+    }
+}