+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Added
+- Support for the inventory/quantity Retail Order Management Public API operation
+
 ## [1.2.0-alpha-2] - 2015-05-21
 ### Added
 - Support for the order/cancel Retail Order Management Public API operation